dfu_load_file: Allow heuristic probe for suffix
If there is no sign of a suffix (CRC code and DFU marker)
let the firmware file through with wildcard match parameters
and big fat warnings!

In the future we might require a suffix on all files, but for now
we will stay as permissive as we have been in previous releases.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>